Shrink MovePicker
Brought sizeof(MovePicker) from 5,480 bytes to 1,896 bytes

Bench 3497901

+++ src/bitboard.h
@@ -15,7 +15,7 @@ inline Square popLsb(Bitboard& bbPtr) {
 }
 
 inline Bitboard getLsb_bb(Bitboard bb) {
-  return bb & (-int64_t(bb));
+  return bb & -bb;
 }
 
 inline Bitboard squareBB(Square sq) {
